refactor(domains): move address-defaults to domains/auth (#232)
